MVP/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
flask==3.0.1
flask-wtf==1.2.1
wtforms==3.0.1
jsonify==0.5
